git-svn-id: svn+ssh://software-srn.sandia.gov/svn/mapreduce/trunk@871 3fc257bb-8550-0410-81b3-a5c103da83f7

+++ oink/rmat2.cpp
@@ -48,8 +48,8 @@ void RMAT2::run()
   // loop until desired number of unique nonzero entries
 
   int niterate = 0;
-  int ntotal = rmat.order * rmat.nnonzero;
-  int nremain = ntotal;
+  uint64_t ntotal = rmat.order * rmat.nnonzero;
+  uint64_t nremain = ntotal;
 
   while (nremain) {
     niterate++;
